requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
joblib==1.4.2
numpy==1.24.3
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.3.0
scipy==1.14.0
six==1.16.0
sqlparse==0.5.1
threadpoolctl==3.5.0
tzdata==2024.1
